+++ crates/externref-xform/src/lib.rs
@@ -101,6 +101,9 @@ impl Context {
     /// large the function table is so we know what indexes to hand out when
     /// we're appending entries.
     pub fn prepare(&mut self, module: &mut Module) -> Result<(), Error> {
+        // Insert reference types to the target features section.
+        wasm_bindgen_wasm_conventions::insert_target_feature(module, "reference-types");
+
         // Figure out what the maximum index of functions pointers are. We'll
         // be adding new entries to the function table later (maybe) so
         // precalculate this ahead of time.