+++ crates/multi-value-xform/src/lib.rs
@@ -117,6 +117,9 @@ pub fn run(
     shadow_stack_pointer: walrus::GlobalId,
     to_xform: &[(walrus::FunctionId, usize, Vec<walrus::ValType>)],
 ) -> Result<Vec<walrus::FunctionId>, anyhow::Error> {
+    // Insert multi-value to the target features section.
+    wasm_bindgen_wasm_conventions::insert_target_feature(module, "multivalue");
+
     let mut wrappers = Vec::new();
     for (func, return_pointer_index, results) in to_xform {
         wrappers.push(xform_one(